Do not archive if already archived

+++ bin/archive_dailynote.py
@@ -2,8 +2,8 @@
 
 from datetime import date
 from datetime import timedelta
-import shutil
 import os
+import shutil
 
 today = date.today()
 yesterday = today - timedelta(days=1)
@@ -11,4 +11,8 @@ home = os.environ['HOME']
 dailynote = home + '/Sync/pkb/main/quicknote.md'
 archivenote = home + '/Sync/pkb/main/diary/' + str(yesterday) + '.md'
 
-shutil.copyfile(dailynote, archivenote)
+# only archive if not already done
+if os.path.exists(archivenote):
+    print(archivenote + ' already exists')
+else:
+    shutil.copyfile(dailynote, archivenote)
